==> 02.Classify.ipynb <==
import pickle
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import chi2
import numpy as np

df = pd.read_excel("learning.xlsx",index_col = 0)
print(len(df))
df["Title"] = df["Title"].astype(str)
df["Description"] = df["Description"].astype(str)
df["StrPrimaryCategoryID"] = df["PrimaryCategoryID"].apply(lambda x: "category"+str(x))
df['Title'] = df[['StrPrimaryCategoryID','Title']].apply(lambda x: ' '.join(x), axis=1)
df['Content'] = df[['Title', 'Description','StrPrimaryCategoryID']].apply(lambda x: ' '.join(x), axis=1)
df.head()

df['Title_Clean'] = df['Title'].str.replace("\r", " ")
df['Title_Clean'] = df['Title_Clean'].str.replace("\n", " ")
df['Title_Clean'] = df['Title_Clean'].str.replace("    ", " ")
df['Title_Clean'] = df['Title_Clean'].str.replace('"', '')
df['Title_Clean'] = df['Title_Clean'].str.lower()

punctuation_signs = list('?:!,;-/*"')

for punct_sign in punctuation_signs:
    df['Title_Clean'] = df['Title_Clean'].str.replace(punct_sign, '')
    
df['Title_Clean'] = df['Title_Clean'].str.replace("'s", "")

# Downloading punkt and wordnet from NLTK
nltk.download('punkt')
print("------------------------------------------------------------")
nltk.download('wordnet')
# Saving the lemmatizer into an object
wordnet_lemmatizer = WordNetLemmatizer()

nrows = len(df)
lemmatized_text_list = []

for row in range(0, nrows):
    
    # Create an empty list containing lemmatized words
    lemmatized_list = []
    
    # Save the text and its words into an object
    text = df.loc[row]['Title_Clean']
    text_words = text.split(" ")

    # Iterate through every word to lemmatize
    for word in text_words:
        lemmatized_list.append(wordnet_lemmatizer.lemmatize(word, pos="v"))
        
    # Join the list
    lemmatized_text = " ".join(lemmatized_list)
    
    # Append to the list containing the texts
    lemmatized_text_list.append(lemmatized_text)

df['Title_Clean'] = lemmatized_text_list
df['Title_Clean']

# Downloading the stop words list
nltk.download('stopwords')
# Loading the stop words in english
stop_words = list(stopwords.words('english'))

# Downloading the stop words list
nltk.download('stopwords')
# Loading the stop words in english
stop_words = list(stopwords.words('english'))

from tqdm.notebook import trange, tqdm

for stop_word in tqdm(stop_words, 
                         total=len(stop_words), 
                         desc="Processing records"):
    regex_stopword = r"\b" + stop_word + r"\b"
    df['Title_Clean'] = df['Title_Clean'].str.replace(regex_stopword, '')

category_codes = {}
category_names = {}
i = 0
for k in df.group.unique().tolist():
    category_codes[k] = i
    category_names[i] = k
    i += 1

df['Category_Code'] = df['group']
df = df.replace({'Category_Code':category_codes})
df.head()

# Going in the fun

DF = df[~(df.group == "None")] 
X_train, X_test, y_train, y_test = train_test_split(df['Title_Clean'], 
                                                    df['Category_Code'], 
                                                    test_size=0.15, 
                                                    random_state=8)

# Parameter election
ngram_range = (1,2)
min_df = 10
max_df = 1.
max_features = 300

ngram_range = (1,2)
min_df = 10
max_df = 1.
max_features = 300

tfidf = TfidfVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,
                        norm='l2',
                        sublinear_tf=True)
                        
features_train = tfidf.fit_transform(X_train).toarray()
labels_train = y_train
print(features_train.shape)

features_test = tfidf.transform(X_test).toarray()
labels_test = y_test
print(features_test.shape)

from sklearn.feature_selection import chi2
import numpy as np

for Product, category_id in sorted(category_codes.items()):
    features_chi2 = chi2(features_train, labels_train == category_id)
    indices = np.argsort(features_chi2[0])
    feature_names = np.array(tfidf.get_feature_names())[indices]
    unigrams = [v for v in feature_names if len(v.split(' ')) == 1]
    bigrams = [v for v in feature_names if len(v.split(' ')) == 2]
    print("# '{}' category:".format(Product))
    print("  . Most correlated unigrams:\n. {}".format('\n. '.join(unigrams[-5:])))
    print("  . Most correlated bigrams:\n. {}".format('\n. '.join(bigrams[-2:])))
    print("")

if 1:
    pathPickle = "Pickles"
else:
    pathPickle = "Pickles_FullDesc"
    
with open(pathPickle+'/df.pickle', 'wb') as output:
    pickle.dump(df, output)
    
with open(pathPickle+'/X_train.pickle', 'wb') as output:
    pickle.dump(X_train, output)
    
# X_test    
with open(pathPickle+'/X_test.pickle', 'wb') as output:
    pickle.dump(X_test, output)
    
# y_train
with open(pathPickle+'/y_train.pickle', 'wb') as output:
    pickle.dump(y_train, output)
    
# y_test
with open(pathPickle+'/y_test.pickle', 'wb') as output:
    pickle.dump(y_test, output)
    
# df
with open(pathPickle+'/df.pickle', 'wb') as output:
    pickle.dump(df, output)
    
    
# features_train
with open(pathPickle+'/features_train.pickle', 'wb') as output:
    pickle.dump(features_train, output)

# labels_train
with open(pathPickle+'/labels_train.pickle', 'wb') as output:
    pickle.dump(labels_train, output)

# features_test
with open(pathPickle+'/features_test.pickle', 'wb') as output:
    pickle.dump(features_test, output)

# labels_test
with open(pathPickle+'/labels_test.pickle', 'wb') as output:
    pickle.dump(labels_test, output)
    
# TF-IDF object
with open(pathPickle+'/tfidf.pickle', 'wb') as output:
    pickle.dump(tfidf, output)
    
    
with open(pathPickle+'/category_codes.pickle', 'wb') as output:
    pickle.dump(category_codes, output)
