==> 09.ComparingModels.ipynb <==
import pickle
import pandas as pd


from pathlib import Path
import glob, os
import os.path


list_pickles = [os.path.join(dp, f) for dp, dn, filenames in os.walk("./Models/") for f in filenames if 'df_' in f]
list_pickles


# https://github.com/pemagrg1/text-classification

df_summary = pd.DataFrame()

for pickle_ in list_pickles:
    
    path =  pickle_
    
    with open(path, 'rb') as data:
        df = pickle.load(data)

    df_summary = df_summary.append(df)

df_summary = df_summary.reset_index().drop('index', axis=1)

df_summary.sort_values('Test Set Accuracy', ascending=False)
